requirements.txt
# Google Genai SDK - pinned compatible version
google-genai==1.5.0

# Vertex AI (for embeddings and other services)
google-cloud-aiplatform>=1.38.0

# Google Auth
google-auth>=2.14.1,<3.0.0

# Numerical operations
numpy>=1.21.0

# Retry logic
tenacity==9.1.2

# Web framework - upgraded for websockets compatibility
gradio==4.44.1

# HTTP requests
requests==2.32.5

# Environment variables
python-dotenv==1.2.1

# Terminal UI
rich==14.2.0
prompt-toolkit==3.0.52

# Utils
typing-extensions==4.15.0

# UUID support (usually built-in but just in case)
uuid

# Python package metadata compatibility for Python 3.9
importlib-metadata>=6.0

# Caching layer
redis==5.0.1
hiredis==2.3.2  # Fast C parser for Redis

# Async support
asyncio
aiohttp==3.9.1
